fixed bug around cookie verification
