+++ internal/application/userservice/user_service.go
@@ -2,7 +2,6 @@ package userservice
 
 import (
 	"context"
-	"errors"
 
 	"filmPackager/internal/domain/project"
 	"filmPackager/internal/domain/user"
@@ -19,61 +18,6 @@ type UserService struct {
 
 func NewUserService(userRepo user.UserRepository, projRepo project.ProjectRepository) *UserService {
 	return &UserService{userRepo: userRepo, projRepo: projRepo}
-}
-
-func (s *UserService) UserLogin(ctx context.Context, email, password string) (*user.User, error) {
-	if email == "" || password == "" {
-		return nil, user.ErrMissingLoginField
-	}
-
-	existingUser, err := s.userRepo.GetUserByEmail(ctx, email)
-	if err != nil {
-		if errors.Is(err, user.ErrUserNotFound) {
-			return nil, user.ErrUserNotFound
-		} else {
-			return nil, fmt.Errorf("error checking for existing user: %v", err)
-		}
-	}
-
-	err = bcrypt.CompareHashAndPassword([]byte(existingUser.Password), []byte(password))
-	if err != nil {
-		return nil, user.ErrInvalidPassword
-	}
-
-	return existingUser, nil
-}
-
-func (s *UserService) CreateUser(ctx context.Context, firstName, lastName, email, password, secondPassword string) (*user.User, error) {
-	// see user_utils.go
-	err := verifyCreateAccountFields(firstName, lastName, email, password, secondPassword)
-	if err != nil {
-		return nil, err
-	}
-
-	username := fmt.Sprintf("%s %s", firstName, lastName)
-
-	hash, err := bcrypt.GenerateFromPassword([]byte(password), bcrypt.DefaultCost)
-	if err != nil {
-		return nil, fmt.Errorf("error hashing password: %v", err)
-	}
-
-	hashedStr := string(hash)
-
-	existingUser, err := s.userRepo.GetUserByEmail(ctx, email)
-	if err != nil && !errors.Is(err, user.ErrUserNotFound) {
-		return nil, user.ErrUserAlreadyExists
-	}
-	if existingUser != nil {
-		return nil, user.ErrUserAlreadyExists
-	}
-
-	newUser := user.CreateNewUser(username, email, hashedStr)
-	err = s.userRepo.CreateNewUser(ctx, newUser)
-	if err != nil {
-		return nil, fmt.Errorf("error creating user: %v", err)
-	}
-
-	return newUser, nil
 }
 
 func (s *UserService) VerifyOldPassword(ctx context.Context, userID uuid.UUID, pw1, pw2 string) error {
